kubernetes_01/requirements.txt
Django==2.0.1
djangorestframework==3.7.7
ipython==6.2.1
python-dotenv==0.8.2
requests==2.18.4
gunicorn==19.7.1
